Create GoalSheet04.md
Made sure to have links to the Pair Programming page, the RPG requirements page, and the RPGDemo repo.  Made sure the maze picture was inline.  Decided on a style for notes.  Square brackets.  Note bold.  From note to the end in italics.  Will need to go back and fix previous goal sheets to reflect that style.