Create CompetencyTests.md
Good ole HTML to the rescue on lists.  Still did not get the interior list to be letters, but that is okay.  And did not realize I would have so many things that MD does not take care of.  (Several lists within lists, some of which did not make sense to use the HTML li.)  Put in horizontal rule (hr) between each competency test.  Breaks it up a bit.  Would like to add links and bookmarks (links on the page), but I am not sure it will work right with the md formatting the headings.  Perhaps later.